+++ bpm-engine-impl/src/main/java/io/takari/bpm/DefaultExecutor.java
@@ -7,6 +7,7 @@ import io.takari.bpm.event.EventPersistenceManager;
 import io.takari.bpm.persistence.PersistenceManager;
 import io.takari.bpm.reducers.*;
 import io.takari.bpm.state.ProcessInstance;
+import io.takari.bpm.task.UserTaskHandler;
 
 import java.util.List;
 import java.util.concurrent.ExecutorService;
@@ -18,7 +19,7 @@ public class DefaultExecutor implements Executor {
     public DefaultExecutor(Configuration cfg, ExpressionManager expressionManager, ExecutorService executor,
                            ExecutionInterceptorHolder interceptors, IndexedProcessDefinitionProvider definitionProvider,
                            UuidGenerator uuidGenerator, EventPersistenceManager eventManager,
-                           PersistenceManager persistenceManager) {
+                           PersistenceManager persistenceManager, UserTaskHandler userTaskHandler) {
 
         this.reducer = new CombiningReducer(
                 new ForkReducer(expressionManager),
@@ -36,7 +37,8 @@ public class DefaultExecutor implements Executor {
                 new ActivationsReducer(interceptors),
                 new FlowListenerReducer(expressionManager),
                 new ScopeReducer(uuidGenerator),
-                new EventGatewayReducer());
+                new EventGatewayReducer(),
+                new UserTaskReducer(userTaskHandler));
     }
 
     @Override