+++ bpm-engine-impl/src/main/java/io/takari/bpm/elements/IntermediateCatchEventHandler.java
@@ -1,10 +1,13 @@
 package io.takari.bpm.elements;
 
+import io.takari.bpm.ProcessDefinitionUtils;
 import io.takari.bpm.actions.Action;
 import io.takari.bpm.actions.CreateEventAction;
 import io.takari.bpm.actions.PopCommandAction;
 import io.takari.bpm.api.ExecutionException;
 import io.takari.bpm.commands.ProcessElementCommand;
+import io.takari.bpm.model.IntermediateCatchEvent;
+import io.takari.bpm.model.ProcessDefinition;
 import io.takari.bpm.state.ProcessInstance;
 
 import java.util.List;
@@ -14,7 +17,11 @@ public class IntermediateCatchEventHandler implements ElementHandler {
     @Override
     public List<Action> handle(ProcessInstance state, ProcessElementCommand cmd, List<Action> actions) throws ExecutionException {
         actions.add(new PopCommandAction());
-        actions.add(new CreateEventAction(cmd.getDefinitionId(), cmd.getElementId()));
+
+        ProcessDefinition pd = state.getDefinition(cmd.getDefinitionId());
+        IntermediateCatchEvent ev = (IntermediateCatchEvent) ProcessDefinitionUtils.findElement(pd, cmd.getElementId());
+        actions.add(new CreateEventAction(cmd.getDefinitionId(), ev));
+
         return actions;
     }
 }