+++ bpm-engine-impl/src/test/java/io/takari/bpm/AbstractEngineTest.java
@@ -1,13 +1,11 @@
 package io.takari.bpm;
 
 import io.takari.bpm.api.BpmnError;
-import io.takari.bpm.api.ExecutionException;
-import org.junit.Before;
-
 import io.takari.bpm.api.Engine;
 import io.takari.bpm.event.EventPersistenceManager;
 import io.takari.bpm.model.ProcessDefinition;
 import io.takari.bpm.task.ServiceTaskRegistryImpl;
+import org.junit.Before;
 
 import java.util.Map;
 import java.util.UUID;
@@ -18,7 +16,7 @@ import static junit.framework.Assert.assertNotNull;
 public class AbstractEngineTest {
 
     private EngineHolder engineHolder;
-    
+
     @Before
     public void init() throws Exception {
         this.engineHolder = new EngineHolder();
@@ -31,15 +29,15 @@ public class AbstractEngineTest {
     protected Configuration getConfiguration() {
         return engineHolder.getConfiguration();
     }
-    
+
     protected Engine getEngine() {
         return engineHolder.getEngine();
     }
-    
+
     protected ServiceTaskRegistryImpl getServiceTaskRegistry() {
         return engineHolder.getServiceTaskRegistry();
     }
-    
+
     protected EventPersistenceManager getEventManager() {
         return engineHolder.getEventManager();
     }
@@ -47,7 +45,11 @@ public class AbstractEngineTest {
     protected ProcessDefinitionProvider getProcessDefinitionProvider() {
         return engineHolder.getProcessDefinitionProvider();
     }
-    
+
+    protected DelegatingUserTaskHandler getUserTaskHandler() {
+        return engineHolder.getUserTaskHandler();
+    }
+
     protected void deploy(ProcessDefinition pd) {
         engineHolder.deploy(pd);
     }
@@ -55,15 +57,15 @@ public class AbstractEngineTest {
     protected void deploy(Map<String, ProcessDefinition> pds) {
         engineHolder.deploy(pds);
     }
-    
-    protected void assertActivations(String processBusinessKey, String processDefinitionId, String ... elementIds) {
+
+    protected void assertActivations(String processBusinessKey, String processDefinitionId, String... elementIds) {
         engineHolder.assertActivations(processBusinessKey, processDefinitionId, elementIds);
     }
 
     protected void dumpActivations() {
         engineHolder.dumpActivations();
     }
-    
+
     protected void assertNoMoreActivations() {
         engineHolder.assertNoMoreActivations();
     }