+++ bpm-engine-impl/src/test/java/io/takari/bpm/EngineHolder.java
@@ -16,6 +16,7 @@ import io.takari.bpm.persistence.PersistenceManager;
 import io.takari.bpm.planner.DefaultPlanner;
 import io.takari.bpm.planner.Planner;
 import io.takari.bpm.task.ServiceTaskRegistryImpl;
+import io.takari.bpm.task.UserTaskHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -46,6 +47,7 @@ public class EngineHolder {
     private final UuidGenerator uuidGenerator;
     private final LockManager lockManager;
     private final Configuration configuration;
+    private final DelegatingUserTaskHandler userTaskHandler;
 
     public EngineHolder() throws Exception {
         serviceTaskRegistry = new ServiceTaskRegistryImpl();
@@ -58,8 +60,10 @@ public class EngineHolder {
         uuidGenerator = new TestUuidGenerator();
         configuration = new Configuration();
         lockManager = new SingleLockManagerImpl();
+        userTaskHandler = new DelegatingUserTaskHandler();
         executor = wrap(new DefaultExecutor(configuration, expressionManager, Executors.newCachedThreadPool(),
-                interceptorHolder, indexedProcessDefinitionProvider, uuidGenerator, eventManager, persistenceManager));
+                interceptorHolder, indexedProcessDefinitionProvider, uuidGenerator, eventManager, persistenceManager,
+                userTaskHandler));
 
         engine = new AbstractEngine() {
 
@@ -115,6 +119,11 @@ public class EngineHolder {
             public EventService getEventService() {
                 return eventService;
             }
+
+            @Override
+            protected UserTaskHandler getUserTaskHandler() {
+                return userTaskHandler;
+            }
         };
 
         activations = new HashMap<>();
@@ -151,6 +160,10 @@ public class EngineHolder {
 
     public ProcessDefinitionProvider getProcessDefinitionProvider() {
         return processDefinitionProvider;
+    }
+
+    public DelegatingUserTaskHandler getUserTaskHandler() {
+        return userTaskHandler;
     }
 
     public void deploy(ProcessDefinition pd) {